requirements.txt
fastapi
uvicorn
requests
azure-cosmos
